+++ machinery/_helpers/__init__.py
@@ -1,0 +1,3 @@
+from . import memoized_property
+
+__all__ = ["memoized_property"]